factored RascalFileSystemServices into an interface with default methods
